Correção de bugs

+++ shortcut-game/src/main/java/com/startupvalley/shortcut_game/game/GameService.java
@@ -1,5 +1,6 @@
 package com.startupvalley.shortcut_game.game;
 
+import com.startupvalley.shortcut_game.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +13,8 @@ public class GameService {
     @Autowired
     private GameRepository gameRepository;
 
+    private static final String ERROR_MESSAGE_NOT_FOUND = "Game not found with id: ";
+
     public List<Game> findAll() {
         return gameRepository.findAll();
     }
@@ -20,11 +23,18 @@ public class GameService {
         return gameRepository.findById(id);
     }
 
-    public Game save(Game game) {
-        return gameRepository.save(game);
+    public void save(Game game) {
+        gameRepository.save(game);
     }
 
     public void deleteById(Long id) {
+        if (!gameRepository.existsById(id)) {
+            throw new ResourceNotFoundException(ERROR_MESSAGE_NOT_FOUND + id);
+        }
         gameRepository.deleteById(id);
+    }
+
+    public String getErrorMessageNotFound(Long id) {
+        return ERROR_MESSAGE_NOT_FOUND + id;
     }
 }
